+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+- The behavior of project to ignore the filename option specified for this plugin in the parent project's gradle properties by default.
+  - For example, if `dotenv.filename=.env.staging` is set in the root project, this setting will automatically apply to sub-projects as well. While this follows the correct resolution order of Gradle Properties, it has been a source of confusion for users working with dotenv.
+  - To disable this default behavior, add `dotenv.filename.ignore.parent=false` to the gradle.properties in the root project.
+  - A same update has been applied to the specification of template file names. To disable this default behavior, add `dotenv.template.filename.ignore.parent=false` to the gradle.properties in the root project.
 
 ### Deprecated